refactor(test): Consolidate test container usage, optimize SQL operations, and centralize container management
Consolidated various test container usage scenarios, optimized SQL operations, and centralized control over container enablement/disabling. This allows seamless testing on local developer machines with automated test container configurations, while enabling Kubernetes-based execution with container management via environment variables.
