+++ src/main/resources/de-init.sql
@@ -1,0 +1,1 @@
+drop table if exists users;
