+++ src/test/java/config/ContainersConfig.java
@@ -1,4 +1,4 @@
-package configuration;
+package config;
 
 import org.springframework.boot.test.context.TestConfiguration;
 import org.springframework.boot.testcontainers.service.connection.ServiceConnection;
