+++ src/test/java/config/ContainersEnabledCondition.java
@@ -1,4 +1,4 @@
-package configuration;
+package config;
 
 import org.springframework.context.annotation.Condition;
 import org.springframework.context.annotation.ConditionContext;
