--- src/test/java/config/ContainersEnvironment.java
@@ -1,15 +1,0 @@
-package config;
-
-import containers.PostgresTestContainer;
-import org.springframework.boot.testcontainers.service.connection.ServiceConnection;
-import org.testcontainers.containers.PostgreSQLContainer;
-import org.testcontainers.junit.jupiter.Container;
-import org.testcontainers.junit.jupiter.Testcontainers;
-
-@Testcontainers
-public class ContainersEnvironment {
-
-    @Container
-    @ServiceConnection
-    public static PostgreSQLContainer postgreSQLContainer = PostgresTestContainer.getInstance();
-}
