--- src/test/java/containers/PostgresTestContainer.java
@@ -1,39 +1,0 @@
-package containers;
-
-import org.testcontainers.containers.PostgreSQLContainer;
-
-public class PostgresTestContainer extends PostgreSQLContainer<PostgresTestContainer> {
-
-    public static final String IMAGE_VERSION = "postgres:15.2-alpine";
-    public static final String USERNAME = "postgres";
-    public static final String PASSWORD = "postgres";
-    public static final String DATABASE_NAME = "tenant1";
-    public static PostgreSQLContainer container;
-
-    public PostgresTestContainer() {
-        super(IMAGE_VERSION);
-    }
-
-    public static PostgreSQLContainer getInstance() {
-        if (container == null) {
-            container = new PostgresTestContainer()
-                    .withDatabaseName(DATABASE_NAME)
-                    .withUsername(USERNAME)
-                    .withPassword(PASSWORD);
-
-        }
-        return container;
-    }
-
-    @Override
-    public void start() {
-        super.start();
-        System.setProperty("DB_URL", container.getJdbcUrl());
-        System.setProperty("DB_USERNAME", container.getUsername());
-        System.setProperty("DB_PASSWORD", container.getPassword());
-    }
-
-    @Override
-    public void stop() {
-    }
-}
